Initial Commit
This is a Wan 2.1 MiniMax Remover workflow.
This is for my personal use, but someone requested I released it, so I am.
There are only simple notes and its not clean nor tidy, but it should be functional.
Also, very recently, there was an issue that caused this to not work in the latest version, so I needed to downgrade to 1.3.0 for it to work. This has reportedly been fixed.